+++ anchore_engine/services/apiext/api/controllers/images.py
@@ -27,234 +27,13 @@ from anchore_engine.db.entities.common import anchore_now
 from anchore_engine.services.apiext.api.controllers.utils import (
     normalize_image_add_source,
     validate_image_add_source,
+    make_response_vulnerability,
 )
 from anchore_engine.subsys import taskstate, logger
 from anchore_engine.subsys.metrics import flask_metrics
 from anchore_engine.utils import parse_dockerimage_string
 
 authorizer = get_authorizer()
-
-
-def make_cvss_scores(metrics):
-    """
-     [
-        {
-          "cvss_v2": {
-            "base_metrics": {
-              ...
-            },
-            "vector_string": "AV:N/AC:L/Au:N/C:P/I:P/A:P",
-            "version": "2.0"
-          },
-          "cvss_v3": {
-            "base_metrics": {
-             ...
-            },
-            "vector_string": "CVSS:3.0/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H",
-            "version": "3.0"
-          },
-          "id": "CVE-2019-1234"
-        },
-        {
-          "cvss_v2": {
-            "base_metrics": {
-              ...
-            },
-            "vector_string": "AV:N/AC:L/Au:N/C:P/I:P/A:P",
-            "version": "2.0"
-          },
-          "cvss_v3": {
-            "base_metrics": {
-             ...
-            },
-            "vector_string": "CVSS:3.0/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H",
-            "version": "3.0"
-          },
-          "id": "CVE-2019-3134"
-        },
-     ]
-    :param metrics:
-    :return:
-    """
-    score_list = []
-
-    for metric in metrics:
-        new_score_packet = {
-            "id": metric.get("id"),
-        }
-        score_list.append(new_score_packet)
-
-        for i in [3, 2]:
-            cvss_dict = metric.get("cvss_v{}".format(i), {})
-            base_metrics = cvss_dict.get("base_metrics", {}) if cvss_dict else {}
-
-            tmp = base_metrics.get("base_score", -1.0)
-            base_score = float(tmp) if tmp else -1.0
-            tmp = base_metrics.get("exploitability_score", -1.0)
-            exploitability_score = float(tmp) if tmp else -1.0
-            tmp = base_metrics.get("impact_score", -1.0)
-            impact_score = float(tmp) if tmp else -1.0
-
-            new_score_packet["cvss_v{}".format(i)] = {
-                "base_score": base_score,
-                "exploitability_score": exploitability_score,
-                "impact_score": impact_score,
-            }
-
-    return score_list
-
-
-def make_response_vulnerability(vulnerability_type, vulnerability_data):
-    ret = []
-
-    if not vulnerability_data:
-        logger.warn("empty query data given to format - returning empty result")
-        return ret
-
-    eltemplate = {
-        "vuln": "None",
-        "severity": "None",
-        "url": "None",
-        "fix": "None",
-        "package": "None",
-        "package_name": "None",
-        "package_version": "None",
-        "package_type": "None",
-        "package_cpe": "None",
-        "package_cpe23": "None",
-        "package_path": "None",
-        "feed": "None",
-        "feed_group": "None",
-        "nvd_data": "None",
-        "vendor_data": "None",
-    }
-
-    osvulns = []
-    nonosvulns = []
-
-    keymap = {
-        "vuln": "CVE_ID",
-        "severity": "Severity",
-        "package": "Vulnerable_Package",
-        "fix": "Fix_Available",
-        "url": "URL",
-        "package_type": "Package_Type",
-        "feed": "Feed",
-        "feed_group": "Feed_Group",
-        "package_name": "Package_Name",
-        "package_path": "Package_Path",
-        "package_version": "Package_Version",
-    }
-    scan_result = vulnerability_data["legacy_report"]
-    try:
-        for imageId in list(scan_result.keys()):
-            header = scan_result[imageId]["result"]["header"]
-            rows = scan_result[imageId]["result"]["rows"]
-            for row in rows:
-                el = {}
-                el.update(eltemplate)
-                for k in list(keymap.keys()):
-                    try:
-                        el[k] = row[header.index(keymap[k])]
-                    except:
-                        el[k] = "None"
-
-                    # conversions
-                    if el[k] == "N/A":
-                        el[k] = "None"
-
-                if el["package_type"].lower() in anchore_engine.common.os_package_types:
-                    osvulns.append(el)
-                else:
-                    nonosvulns.append(el)
-
-                el["nvd_data"] = []
-                el["vendor_data"] = []
-                if row[header.index("CVES")]:
-                    all_data = json.loads(
-                        row[header.index("CVES")]
-                    )  # {'nvd_data': [], 'vendor_data': []}
-                    el["nvd_data"] = make_cvss_scores(all_data.get("nvd_data", []))
-                    el["vendor_data"] = make_cvss_scores(
-                        all_data.get("vendor_data", [])
-                    )
-
-    except Exception as err:
-        logger.exception("could not prepare query response")
-        logger.warn("could not prepare query response - exception: " + str(err))
-        ret = []
-
-    # non-os CPE search
-    keymap = {
-        "vuln": "vulnerability_id",
-        "severity": "severity",
-        "package_name": "name",
-        "package_version": "version",
-        "package_path": "pkg_path",
-        "package_type": "pkg_type",
-        "package_cpe": "cpe",
-        "package_cpe23": "cpe23",
-        "url": "link",
-        "feed": "feed_name",
-        "feed_group": "feed_namespace",
-    }
-    scan_result = vulnerability_data["cpe_report"]
-    # gather a map of package path->CVE IDs referred by vulndb entries, dedup logic copied from policy engine
-    dedup_hash = {}
-    for vuln in scan_result:
-        if vuln.get("feed_name") == "vulndb":
-            pkg_path = vuln.get("pkg_path")
-            for nvd_item in vuln.get("nvd_data", []):
-                if pkg_path not in dedup_hash:
-                    dedup_hash[pkg_path] = set()
-                dedup_hash[pkg_path].add(nvd_item.get("id"))
-
-    for vuln in scan_result:
-        # dedup pass: skip nvd record if vulndb covers it
-        if dedup_hash and vuln.get("feed_name") != "vulndb":
-            pkg_path = vuln.get("pkg_path")
-            if (
-                pkg_path in dedup_hash
-                and vuln.get("vulnerability_id") in dedup_hash[pkg_path]
-            ):
-                continue
-
-        el = {}
-        el.update(eltemplate)
-
-        for k in list(keymap.keys()):
-            el[k] = vuln[keymap[k]]
-
-        if vuln["name"] != vuln["version"]:
-            pkg_final = "{}-{}".format(vuln["name"], vuln["version"])
-        else:
-            pkg_final = vuln["name"]
-
-        el["package"] = pkg_final
-
-        # get nvd scores
-        el["nvd_data"] = []
-        el["nvd_data"] = make_cvss_scores(vuln.get("nvd_data", []))
-
-        # get vendor scores
-        el["vendor_data"] = []
-        el["vendor_data"] = make_cvss_scores(vuln.get("vendor_data", []))
-
-        fixed_in = vuln.get("fixed_in", [])
-        el["fix"] = ", ".join(fixed_in) if fixed_in else "None"
-
-        nonosvulns.append(el)
-
-    if vulnerability_type == "os":
-        ret = osvulns
-    elif vulnerability_type == "non-os":
-        ret = nonosvulns
-    elif vulnerability_type == "all":
-        ret = osvulns + nonosvulns
-    else:
-        ret = vulnerability_data
-
-    return ret
 
 
 def make_response_policyeval(eval_record, params, catalog_client):
